requirements.txt
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Dance[sqla]==3.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
blinker==1.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
